version 9

+++ Alertix_DB/src/main/java/com/scr/Alertix_DB/Model/Categoria.java
@@ -14,7 +14,7 @@ public class Categoria {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int idCategorias;
+    private Long idCategorias;
     @NotBlank
     private String nombreCategoria;
